+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 1.1.0 - 2016-03-23
+
+### Added
+
+- [#3](https://github.com/zendframework/zend-mvc-console/pull/3) adds the
+  `CreateConsoleNotFoundModel` controller plugin from zend-mvc. This also
+  required adding `Zend\Mvc\Console\Service\ControllerPluginManagerDelegatorFactory`
+  to ensure it is present in the controller plugin manager when in a console
+  context.
+- [#3](https://github.com/zendframework/zend-mvc-console/pull/3) adds
+  `Zend\Mvc\Console\Service\ControllerManagerDelegatorFactory`, to add an
+  initializer for injecting a console adapter into `AbstractConsoleController`
+  instances.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#3](https://github.com/zendframework/zend-mvc-console/pull/3) updates the
+  `AbstractConsoleController` to override the `notFoundAction()` and always
+  return the return value of the `CreateConsoleNotFoundModel` plugin.
+- [#3](https://github.com/zendframework/zend-mvc-console/pull/3) updates the
+  `AbstractConsoleController` to mark it as abstract, as was always intended,
+  but evidently never implemented, in zend-mvc.
+
 ## 1.0.1 - TBD
 
 ### Added